ダークモードのAtom化
