+++ src/components/darkModeButton.tsx
@@ -1,12 +1,27 @@
-import { memo } from 'react';
+import { ChangeEvent, memo, useCallback, useEffect } from 'react';
 import { Stack, Switch } from '@mui/material';
 import { TitleStyle } from '@/utils/styleComponents';
-import { useAtomValue } from 'jotai';
-import { darkModeAtom, useDarkMode } from '@/hooks/useDarkMode';
+import { useAtom } from 'jotai';
+import { darkModeAtom } from '@/globalStates/darkModeAtom';
 
 export const DarkModeButton = memo(() => {
-  const isDarkMode = useAtomValue(darkModeAtom);
-  const { handleDarkMode } = useDarkMode(); // ダークモードを管理するカスタムフック
+  const [isDarkMode, setIsDarkMode] = useAtom(darkModeAtom);
+
+  const handleDarkMode = useCallback((event: ChangeEvent<HTMLInputElement>) => {
+    setIsDarkMode(event.target.checked);
+  }, []);
+
+  // 初回読み込み時にOSのテーマを読み込む
+  useEffect(() => {
+    const prefersColorSchemeDark = matchMedia('(prefers-color-scheme: dark)').matches;
+
+    // ダークテーマに設定されていたらisDarkModeをtrueにする
+    if (prefersColorSchemeDark) {
+      setIsDarkMode(true);
+    } else {
+      setIsDarkMode(false);
+    }
+  }, []);
 
   return (
     <Stack spacing={1}>
