+++ src/globalStates/darkModeAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'jotai';
+
+/**
+ * ダークモードを管理するState
+ */
+export const darkModeAtom = atom(false);
